- Update Tailwind CSS, React & React Dom, NextJs Packages - Create Layouts & Modules Components - Rearrage & Rename static images & SVG files - Added Global.css file
