+++ package.json
@@ -61,12 +61,12 @@
 		"eslint-plugin-simple-import-sort": "^7.0.0",
 		"jest": "^26.5.3",
 		"jest-dom": "^4.0.0",
-		"postcss": "^8.2.10",
+		"postcss": "^8.2.13",
 		"prettier": "^2.0.5",
-		"tailwindcss": "^2.1.1"
+		"tailwindcss": "^2.1.2"
 	},
 	"dependencies": {
-		"@headlessui/react": "^1.0.0",
+		"@headlessui/react": "^1.1.1",
 		"@heroicons/react": "^1.0.1",
 		"@material-ui/core": "^4.9.12",
 		"@material-ui/icons": "^4.9.1",
@@ -88,15 +88,17 @@
 		"mobx": "5.15.4",
 		"mobx-react": "6.3.0",
 		"moment": "^2.28.0",
-		"next": "latest",
+		"next": "^10.2.0",
 		"next-on-netlify": "^2.6.3",
 		"nprogress": "^0.2.0",
 		"parse": "^3.1.0",
+		"postcss-import": "^14.0.1",
+		"postcss-nested": "^5.0.5",
 		"prefs-rcl": "^0.1.8",
 		"prop-types": "15.7.2",
 		"raw-loader": "^4.0.2",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
 		"react-intl": "^4.7.6",
 		"sweetalert": "2.1.2",
 		"winston": "^3.3.3"
